Improved how the method "runInvivoTestingSession" is implemented during testing sessions.
The actual "runInvivoTestingSession" will contain locks and other synchronization
statements that may eventually neglect the execution of the in-vivo testing session.
During the test it is important to be (almost) sure that the runInvivoTestingSession,
is not skipped, otherwise it is likely that the tests become flaky.
The override of the method "runInvivoTestingSession" reduces this risk by trying
to run on in-vivo testing session several times (i.e. MAX_TENTATIVES)